feat(convert): BIG to native refactor + blocks
