+++ bpy_speckle/convert/__init__.py
@@ -1,73 +1,14 @@
 import bpy, idprop, bpy_types
+from bpy_speckle.convert.to_native import CAN_CONVERT_TO_NATIVE, convert_to_native
 from mathutils import Matrix
 from devtools import debug
 
-from .from_speckle import *
 from .to_speckle import *
 from .util import *
 from bpy_speckle.functions import _report, get_scale_length
 
 from specklepy.objects.geometry import *
 from specklepy.objects.other import BlockInstance, RenderMaterial
-
-
-def transform_to_native(transform, scale=1.0):
-    mat = Matrix(
-        [
-            transform.value[:4],
-            transform.value[4:8],
-            transform.value[8:12],
-            transform.value[12:16],
-        ]
-    )
-    # scale the translation
-    for i in range(3):
-        mat[i][3] *= scale
-    return mat
-
-
-def block_def_to_native(definition, scale=1.0):
-    _report(f">>> creating block definition for {definition.name} ({definition.id})")
-    native_def = bpy.data.collections.get(definition.name)
-    if native_def:
-        return native_def
-
-    native_def = bpy.data.collections.new(definition.name)
-    for geo in definition.geometry:
-        b_obj = from_speckle_object(geo, scale)
-        native_def.objects.link(b_obj)
-
-    return native_def
-
-
-def import_block(instance, scale=1.0, name=None):
-    """
-    Convert BlockInstance to native
-    """
-    _report(f">>> converting block instance {instance.id}")
-
-    name = getattr(instance, "name", f"BlockInstance -- {instance.id}")
-    debug(name)
-    native_def = block_def_to_native(instance.blockDefinition, scale)
-
-    native_instance = bpy.data.objects.new(name, None)
-    native_instance.instance_collection = native_def
-    native_instance.instance_type = "COLLECTION"
-    native_instance.matrix_world = transform_to_native(instance.transform, scale)
-
-    return native_instance
-
-
-FROM_SPECKLE_SCHEMAS = {
-    Mesh: import_mesh,
-    Brep: import_brep,
-    Curve: import_curve,
-    Line: import_curve,
-    Polyline: import_curve,
-    Polycurve: import_curve,
-    Arc: import_curve,
-    BlockInstance: import_block,
-}
 
 
 TO_SPECKLE = {
@@ -87,43 +28,9 @@ def set_transform(speckle_object, blender_object):
         transform = speckle_object.properties.get("transform", None)
 
     if transform and len(transform) == 16:
-        mat = Matrix(
-            [transform[0:4], transform[4:8], transform[8:12], transform[12:16]]
-        )
+        mat = Matrix([transform[:4], transform[4:8], transform[8:12], transform[12:16]])
+
         blender_object.matrix_world = mat
-
-
-def add_blender_material(smesh, blender_object) -> None:
-    """Add material to a blender object if the corresponding speckle object has a render material"""
-    if blender_object.data is None:
-        return
-
-    if not hasattr(smesh, "renderMaterial") and not hasattr(smesh, "@renderMaterial"):
-        return
-
-    speckle_mat = getattr(smesh, "renderMaterial", None) or smesh["@renderMaterial"]
-    mat_name = getattr(speckle_mat, "name", None) or speckle_mat.__dict__.get("@name")
-    if not mat_name:
-        mat_name = speckle_mat.applicationId or speckle_mat.id or speckle_mat.get_id()
-    blender_mat = bpy.data.materials.get(mat_name)
-    if not blender_mat:
-        blender_mat = bpy.data.materials.new(mat_name)
-
-        # for now, we're not updating these materials. as per tom's suggestion, we should have a toggle
-        # that enables this as the blender mats will prob be much more complex than whatever is coming in
-        blender_mat.use_nodes = True
-        inputs = blender_mat.node_tree.nodes["Principled BSDF"].inputs
-
-        inputs["Base Color"].default_value = to_rgba(speckle_mat.diffuse)
-        inputs["Emission"].default_value = to_rgba(speckle_mat.emissive)
-        inputs["Roughness"].default_value = speckle_mat.roughness
-        inputs["Metallic"].default_value = speckle_mat.metalness
-        inputs["Alpha"].default_value = speckle_mat.opacity
-
-    if speckle_mat.opacity < 1:
-        blender_mat.blend_method = "BLEND"
-
-    blender_object.data.materials.append(blender_mat)
 
 
 def material_to_speckle(blender_object) -> RenderMaterial:
@@ -176,26 +83,6 @@ def try_add_property(speckle_object, blender_object, prop, prop_name):
 #                 pass
 
 
-def add_custom_properties(speckle_object, blender_object):
-
-    if blender_object is None:
-        return
-
-    blender_object["_speckle_type"] = type(speckle_object).__name__
-    # blender_object['_speckle_name'] = "SpeckleObject"
-
-    ignore = ["_chunkable", "_units", "units"]
-
-    if hasattr(speckle_object, "applicationId"):
-        blender_object["applicationId"] = speckle_object.applicationId
-
-    for key in speckle_object.get_dynamic_member_names():
-        if key in ignore:
-            continue
-        if isinstance(speckle_object[key], (int, str, float, dict)):
-            blender_object[key] = speckle_object[key]
-
-
 def dict_to_speckle_object(data):
     if "type" in data.keys() and data["type"] in SCHEMAS.keys():
         obj = SCHEMAS[data["type"]].parse_obj(data)
@@ -226,43 +113,17 @@ def from_speckle_object(speckle_object, scale, name=None):
         or getattr(speckle_object, "name", None)
         or speckle_object.speckle_type + f" -- {speckle_object.id}"
     )
-
-    units = getattr(speckle_object, "units", None)
-    if units:
-        scale = get_scale_length(units) / bpy.context.scene.unit_settings.scale_length
-
     # try native conversion
-    if type(speckle_object) in FROM_SPECKLE_SCHEMAS.keys():
-        print("Got object type: {}".format(type(speckle_object)))
+    if type(speckle_object) in CAN_CONVERT_TO_NATIVE:
+        print(f"Got object type: f{type(speckle_object)}")
 
         try:
-            obdata = FROM_SPECKLE_SCHEMAS[type(speckle_object)](
-                speckle_object, scale, speckle_name
-            )
+            blender_object = convert_to_native(speckle_object, speckle_name)
         except Exception as e:  # conversion error
             _report(f"Error converting {speckle_object} \n{e}")
+            raise e
             return None
 
-        if speckle_name in bpy.data.objects.keys():
-            blender_object = bpy.data.objects[speckle_name]
-            blender_object.data = obdata
-            blender_object.matrix_world = Matrix()
-            if hasattr(obdata, "materials"):
-                blender_object.data.materials.clear()
-        else:
-            debug(speckle_name)
-            debug(obdata)
-            debug(obdata.name)
-            blender_object = (
-                obdata
-                if isinstance(obdata, bpy_types.Object)
-                else bpy.data.objects.new(speckle_name, obdata)
-            )
-
-        blender_object.speckle.object_id = str(speckle_object.id)
-        blender_object.speckle.enabled = True
-
-        add_custom_properties(speckle_object, blender_object)
         add_blender_material(speckle_object, blender_object)
         # TODO: transforms
         # set_transform(speckle_object, blender_object)
@@ -285,21 +146,18 @@ def from_speckle_object(speckle_object, scale, name=None):
             return from_speckle_object(display, scale, speckle_name)
 
     # return none if fail
-    _report("Invalid input: {}".format(speckle_object))
+    _report(f"Could not convert usupported Speckle object: {speckle_object}")
     return None
 
 
 def get_speckle_subobjects(attr, scale, name):
-
     subobjects = []
     for key in attr.keys():
         if isinstance(attr[key], dict):
             subtype = attr[key].get("type", None)
             if subtype:
                 name = "{}.{}".format(name, key)
-                # print("{} :: {}".format(name, subtype))
                 subobject = from_speckle_object(attr[key], scale, name)
-                add_custom_properties(attr[key], subobject)
 
                 subobjects.append(subobject)
                 props = attr[key].get("properties", None)
@@ -309,9 +167,7 @@ def get_speckle_subobjects(attr, scale, name):
             subtype = attr[key].type
             if subtype:
                 name = "{}.{}".format(name, key)
-                # print("{} :: {}".format(name, subtype))
                 subobject = from_speckle_object(attr[key], scale, name)
-                add_custom_properties(attr[key], subobject)
 
                 subobjects.append(subobject)
                 props = attr[key].get("properties", None)
@@ -320,7 +176,7 @@ def get_speckle_subobjects(attr, scale, name):
     return subobjects
 
 
-ignored_keys = [
+ignored_keys = (
     "speckle",
     "_speckle_type",
     "_speckle_name",
@@ -329,7 +185,7 @@ ignored_keys = [
     "transform",
     "_units",
     "_chunkable",
-]
+)
 
 
 def get_blender_custom_properties(obj, max_depth=1000):
