--- bpy_speckle/convert/from_speckle/__init__.py
@@ -1,3 +1,0 @@
-from .mesh import import_mesh
-from .curve import import_curve
-from .brep import import_brep
