--- bpy_speckle/convert/from_speckle/plane.py
@@ -1,6 +1,0 @@
-import bpy, bmesh, struct
-import base64
-
-
-def import_plane(speckle_plane, scale=1.0, name=None):
-    return None
